refactor: separate headers
